iseditor and ismoderator kernel auth

+++ routes/web.php
@@ -37,11 +37,11 @@ Route::controller(HotelInfoController::class)->group(function () {
 
 Route::controller(UserController::class)->group(function () {
     Route::get('/back/users', 'index')->name("users");
-    Route::get('/back/users/create', 'create')->name("userscreate");
-    Route::get('/back/users/{id}/show', 'show');
-    Route::post('/back/users/create', 'store');
-    Route::put('/back/users/{id}/update', 'update');
-    Route::delete('/back/users/{id}/delete', 'destroy');
+    Route::get('/back/users/create', 'create')->middleware('isAdmin')->name("userscreate");
+    Route::get('/back/users/{id}/show', 'show')->middleware('isAdmin');
+    Route::post('/back/users/create', 'store')->middleware('isAdmin');
+    Route::put('/back/users/{id}/update', 'update')->middleware('isAdmin');
+    Route::delete('/back/users/{id}/delete', 'destroy')->middleware('isAdmin');
 });
 
 //testimonials
